[Telemetry] Add generic typ expectation tags for desktop and mobile platforms
This CL adds the desktop, mobile, android-webview, android-not-webview and
android-low-end tags. These tags can be used to shorten the amount of expectations.
Instead of writing an expectation for each desktop or mobile platform, only one
expectation can be written using a generic tag. For instance the desktop tag can
be used to substitute the linux, mac, windows and chromeos tags.

Bug: chromium:973936
Change-Id: Id7258698154d06b6abb1a6e0ebd3d9a22fec50db