Update patch set 1
Patch Set 1:

Is there an advantage to making this be part of typ, if it's a telemetry-specific thing? 

I.e., given that the Expectation is a public object, why not just do the serialization in story/typ_expectations.py instead?

I worry that putting it in typ might encourage other people to use this and I'm not sure that we want to encourage that ...

Patch-set: 1
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>