Pinpoint - Record job timings
Record job timings on job completion, indexed by the jobs type,
configuration, benchmark, and story. This allows us to query for
historical timings of similar jobs to get an idea of how long a new
job will take. Next steps will be to query the length of the queue
ahead of us, and to surface all this clearly on the job page.

Bug: chromium:987371
Change-Id: I19809b0c2ab201841369283fcd90427668dff724