[Typ] Create a public static method in TestExpectations which finds unused expectation patterns
This method takes a dictionary that maps expectation patterns to lists of expectation instances
and a trie with test names. The method searches for the expectations patterns in the trie of test names
If it does not find the pattern then it puts the expectations it is mapped to in a list of unused
expectations and returns the list.

Bug: chromium:973936
Change-Id: Ibb974e70316f0395c61b6ab65cbc3665eef8c176