[Typ] Expectation tags in tag declaration lists are not case insensitive
This CL will make the tags in the tag declaration lists be stored in lower
case. Then those tags will be case insensitive. This CL also makes sure that
the return value of parse_tagged_list is 1 if the conflicts_allowed descriptor
is set to False and there are conflicts between test expectations.

Bug: chromium:973936
Change-Id: I631cde71014223c3b02fded09ae5a0b7835f4515