Update patch set 3
Patch Set 3: Commit-Queue+1

Thanks both for the comments! I went for just quoting special chars in story names; and not making it sound like this is a "requirement", but more like a temporary workaround to make our lives a little easier with the intermediate results format until crbug.com/983993 is resolved.

> What benefit do we expect from option 2? 
> cost: does this require a migration (busy work)?
> Can the same benefit be gained by establishing that the first component of a slash separate path is benchmark_name, and simply doing
> 
>   benchmark_name, story_name = test_path.split('/', 1)
> 
> ?

Benefit is making it easier to deal with these paths wherever they might eventually show up. Yes, at the moment it's only benchmark_name/story_name, and a simple approach like you suggested works. But in the past test paths like this have ended up showing as e.g. benchmark_name/builder/measurement/[story_group/]story_name on chrome perf or path.to.unittest.Class.benchmark_name/sory_name on smoke tests. And it's often a pain to properly parse and interpret those paths.

Cost is 0, because this is only for new code paths. No migrations needed because we can easily translate these and fall back to the hard-to-work-with-but-that-is-what-they-are-for-now paths when interfacing with all existing systems.

Patch-set: 3
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1