PinpointScheduler: Speculate in Bisection
This change makes Pinpoint do speculative bisection by also setting the
second-level bisection tree in a binary search as potential changes to
explore. This allows us to better utilise the machines we have when
bisecting a range of changes, and also for scheduling the builds for
intermediate changes.

Now that we have the FIFO Scheduler to enforce pseudo-exclusivity on the
pool of machines running the bisection tests, we can better leverage the
Swarming scheduler to then let the work be distributed on available
machines during bisections. For now we're only speculatively exploring
the binary search tree one level down as a stop-gap implementation (in
the future we can raise this to 2 levels to run 7 builds/tests on
revisions instead of 3).

As a visualisation of the approach, here we have an arbirary (linear)
revision range with N internal members:

  [ c0, c1, c2, ..., cN ]

We can then imagine a binary search tree rooted at c[(N - 0) / 2]:

               c[N/2]
             /        \
            /          \
      c[N/4]            c[N 3/4]
     /      \          /        \
  ...        ...    ...          ...

Instead of having to wait for c[N/2] to finish, we're adding both c[N/4]
(which is the midpoint between c[0] and c[N/2]) and c[N 3/4] (which is
the midpoint between c[N/2] and c[N]) which allows us to converge at
potential culprits quicker in terms of wall-time if we're able to
schedule and run all the builds and tests to run for all three of the
candidate culprits in the same amount of time it would take us to wait
on just one change to finish. This allows us to use the available
machines for bisection much better.

We can then use this information, along with queue lengths to determine
how much more machines we'll need to effectively run bisections
regularly to identify regressions on-demand.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:986788
Change-Id: I769ba266461af5267d6dc8964bc48438b08dd01d