[Typ] Refactor get_broken_expectations to accept a list of test names instead of a trie
The creation of the trie should be done within typ.expectations_parser.
Code outside of the expectations_parser module shouldn't need to know how to create
the trie that is passed to get_broken_expectations. This CL will change
the argument into a list of test names, and get_broken_expectations will
create a trie from the test names in the list.

Bug:chromium:973936
Change-Id: I03195de53951220212916f9207efe38a20f7f52b