[Telemetry] Clean up page_test_results tests
Make sure the results object is always used in the context managed
form.

This allows to remove the direct calls to PrintSummary, which will be
removed when all clients use the context managed version.

Also take the chance to remove BaseTestResultsUnittest, there was
only a single client of this class and it wasn't getting much out
of it.

Bug: chromium:973837
Change-Id: Ibc2ecea990055c70f5db176c243769e826337e53