[Telemetry] Set intermediate_dir independent of output_dir
This CL adds an intermediate_dir argument to PageTestResults. This
directory is used for storing intermediate test results, which are then
passed to results_processor. Final results remain in output_dir. This
makes it possible to keep intermediate results of several benchmark
runs with the same output_dir.

The default behavior is the same as before: intermediate results are
stored in {output_dir}/artifacts/.

Bug: chromium:981349
Change-Id: I223abaa9f41754d641336d6b0e8a0273518586ff