Fix how WPR injects scripts on pages with a Content Security Policy where 'default-src' and 'script-src' are both listed as directives.
A few sites are failing now in the captured sites framework because a chrome
update that tightened up the CSP and not have inheritance
See https://www.w3.org/TR/CSP3/ section 6.1.3. default-src

Previously, code looked for 'default-src' or 'script-src' and only updated
whichever it saw first

This change treats the 'default-src' directive as a proper fallback only to be used when 'script-src' is not listed.

This fix unblocks autofill test on indigo.ca and
password tests on airbnb, github, paypal, and discord

Bug: chromium:993561
Change-Id: I88f4f489e060c8391d649045c91d4d77eaff8dc0