Refactor ApkHelper and DeviceUtils to support APKS archives
Previously, we supported Android app bundles by invoking the bundle
install script. This had several drawbacks:
- We couldn't use the common APK operations in ApkHelper.
- We had to special case bundle install in DeviceUtils.
Furthermore, to support the bundle install script in swarming we have to
add the signing key and several other artifacts to isolate.

A better approach is to use APKS archives (a zip file with prebuilt
APKs) as the installable artifact. This CL adds support for APKS
archives by abstracting away the installable artifact in apk_helper.py.
Abstraction is achieved with an abstract base class for an installable
app and subclasses for single APKs, split APKs, APKS archives and also
the bundle install script to keep support current bots.

Change-Id: Icf290ca3e68b5fbddf4b87e78bd389927056cba8
Bug: chromium:989654