Dashboard - Rework update_test_suite_descriptors a bit.
There was a few problems with this endpoint, that investigation in
crbug.com/992162 revealed. First off, it takes forever to update, to
the point that I think the deferred queue was going back days or weeks.
Second, it wasn't properly breaking up test paths into
measurement/story.

I don't know what the plan is for this endpoint, but for the short
term at least this makes the data showing up in Pinpoint correct. I
removed most of the dependencies on descriptors since those were
a v2 abstraction thats we're not using anymore.

Bug: chromium:992162
Change-Id: Iae934a5e115037859fd70ea2f9138241d6793827