Update patch set 1
Patch Set 1:

Thinking a bit more about this, all test results in a single run will have the same shard index. So maybe the right place to put this is on a top level benchmarkRun info dict?

The only reason to keep them as tags is in case we want to query about this info when this is stored on some BQ database. But maybe this would be redundant and already stored elsewhere? (e.g. together with the builder name and other metadata about the test environment?)

Patch-set: 1
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>