Revert "[Telemetry] Cleanup dictionary that maps the old expectations format's tags to the new tags"
This reverts commit 2ae52adeeeddfebecc9f4b1a5236b8cbdb42589c.

Reason for revert: Re add dictionary that maps old expectation tags to
typ's expectations tags. This dictionary can be used to translate the
new format back into the old format and vice versa. We may do this if
we want to temporarily switch back to the old format because there are
bugs in the new format.

Original change's description:
> [Telemetry] Cleanup dictionary that maps the old expectations format's tags to the new tags
> 
> Bug: chromium:973936
> Change-Id: I4bbbb037bf8f806f40dd87d815696aab3aeb0514
> Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1754500
> Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
> Commit-Queue: Rakib Hasan <rmhasan@google.com>

TBR=perezju@chromium.org,rmhasan@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Bug: chromium:973936
Change-Id: I21ed09c61cdce7f65759d116b311e805553fd3f0