[Telemetry] Avoid searching for actual browsers in browser test case unit tests
The call to GenerateTags during browser test initialization requires us to find a
possible browser. However, searching for possible browsers is very expensive and
also can be very flaky . We even search for android devices when we want to only
run the tests on a desktop. Also for actual browser tests like gpu tests, the
GenerateTags function creates a browser session just for getting device tags as well
as platform tags. Therefore it is inefficient to call GenerateTags at browser test
initialization. Also for example browser test cases for unit tests we should use
a fake possible browser, so that we can avoid a very flaky and expensive search for
possible browsers.

Bug:chromium:992260
Change-Id: Ibc0cb62144907c41e1c03fbbb8610a0ebf3408dc