Update patch set 1
Patch Set 1: Commit-Queue+1

Note to reviewers:

- Please review for readability, since there's some complexity in the graph traversal implementation (even thought it's "just a DFS"). In particular we're using graph colouring in the traversal to achieve the post-order evaluation.

- The tests are *long* mostly due to boilerplate. I'm happy to keep those as-is but if you'd like me to iterate more on those, I'd be really happy to oblige.

- This is lacking documentation (purposefully) but I'd really like to at least have module-level and public function-level docstrings describing the semantics. I'm happy to spend as many lines as required in documentation to demystify what this is doing.

I'll be stacking on some dependent changes which attempt to implement the logic we have for the various "quests" we have. Ideally the evaluator can follow an OOP model, since we're using a `Callable` interface for the `evaluator` if that's more helpful than the explicit passing of an accumulator (representing memory).

Also please note that this is as far as I'd like to go towards implementing an interpreter for a domain-specific-language. :)

Patch-set: 1
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1