Update patch set 7
Patch Set 7:

(2 comments)

I noticed something when poking through the demo, that the "start" of the range isn't in the dropdown anymore once you clear it. Or at least it isn't as clear because the commit position isn't shown in the dropdown but is shown in the underlying UI.

Would it be possible to optionally show the commit position as well if it's in the revision, and leave it out if it's not?

Looking at this in particular: https://dev-simonhatch-9044ff96-dot-pinpoint-dot-chromeperf.appspot.com/job/136a27967c0000

Patch-set: 7
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>