Pinpoint: Post issue update on job creation
This makes Pinpoint post an update to the associated bug that a Pinpoint
job has been created and queued. Since we've implemented a scheduler, we
moved the "job started" message later, causing some confusion with users
that were used to seeing the job started messages posted as soon as the
job is created.

Now we're posting an update as soon as we've scheduled the job, which
might race with the scheduler running the job and scheduling it. This is
fine because it's better to provide the message even out-of-order than
it is to leave no trace of the scheduling action being performed.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:998668
Change-Id: I2959ecd51b8a61e4d3084cee63a9c9e8c21c8de1