Pinpoint - Add bisect menu w/ retry+edit
This adds a bisect menu and allows editing of the params. Mostly just
a few new inputs and a radio button. Theres a <commit-details>
dom-module that wraps displaying a git hash + calling out to the
backend for more info on it. It also allows a range of commits to be
inserted, and can display them w/ hash, message, and author.

Bug: chromium:992492
Change-Id: Iaaa9ba4435af37bbe5264860b0da5a685c55de55