Symbolize object type with frame name if enable "--frame-as-object-type"
By default, the "Object type" tab may contain [unknown] or other less
informative data on Cast platform, which will add additional efforts to
demystify the content. not give accurate information about objects. If
"--frame-as-object-type" is enabled, we replace object type with program
counter from the frame that allocated the object and symbolize it.

Bug: chromium:1001848