[metrics] Move filtering of V8 GC metrics to the metrics code
Currently V8 GC metrics are filtered on benchmark level using
ShouldAddValue predicate. That will be removed soon, so we need to
filter in the metric code.

This patch disables metrics for GC phases by default.
It also adds an optional whitelist parameter to groupAndProcessEvents.
Any group that is not in the whitelist will be ignored.
As a side-effect this fixes crbug.com/673311

Bug: chromium:1000187
Change-Id: Idf91d69b55624359bc33cdc01e204ca333e762ef