PinpointScheduler: The Execution Engine Core
This change introduces the Task and TaskLog datastore entity models
which are descendants of the Job model. We use this to implement an
execution engine which relies on the dependency structure of tasks to
express the evaluation of the state of a job. This execution engine
allows us to decouple the handling of how to react to tasks in a
separate module given an event, while concentrating on the evaluation of
the task dependency graph.

We do this by implementing a depth-first traversal of a dependency graph
starting from the terminal vertices (those that don't have any dependent
vertices) which we then apply an evaluator to from the "leaf" vertices
(those that do not depend on any other vertices). This would represent a
post-order traversal (leaves first then immediate parent, bubbling up
the dependency graph) where we apply an evaluator which has access to
the current task being evaluated, the event we're processing (input),
and an accumulator dictionary (output).

The main entry-point for the algorithm is the `Evaluate` function which
hides the details of the graph in the module interface. External
functions only get access to the materialised `TraversalTask` named
tuple instead of the actual NDB model for the `Task`. In lieu of direct
access to the model object we provide transactional graph-modification
and `TaskLog` appending available through the module interface. This
lets us implement the evaluators independently of the traversal through
inversion of control.

This code is isolated for now to allow us to test in isolation. Future
changes will implement the event-handling specific logic that parallels
the existing modules in the `pinpoint.model.quest` submodule, and then
integrate with the PubSub handler. This approach allows us to
incrementally deploy the event-based Pinpoint job handling without
disrupting the current pull-based implementation.

Eventually we intend to do a randomised trial to determine whether a
Pinpoint job will be relying on the polling-based implementation or the
event-based implementation.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: Ie206d82b8c4061a135599089c2cf056f2c9e9264