[Telemetry] Check if a story can be run before launching the browser
Currently we check if a story can be actually run only after the browser
has launched and the tracing has been started. This is a waste of time,
and also it will make Results Processor to try to compute metrics on the
produced traces.

This CL moves the check up, so that we could skip the story before the
browser is launched.

Bug: chromium:1002272
Change-Id: Ia66ce694163ebafd9f36f2ca7e156560e1675dbe