PinpointScheduler: Port of find_isolate to Tasks
This change implements the evaluator for the `find_isolate` quest in
Pinpoint to work with the execution engine core. We start with the
`find_isolate` quest because it's typically a leaf task where we can
evaluate the state of the task in isolation. This gives us an example of
how to compose evaluators, and how we can coordinate the event structure
with the specific evaluators.

This decomposition also allows us to structure the code to be reactive
with smaller, well-defined actions that we can test in isolation. The
actions can perform not just state transformations in a transactional
manner, we can also get a "global" picture of the accumulator from the
actions after a full traversal. This allows us to make decisions on
whether we should continue an action based on the overall state of the
accumulator.

Bug: chromium:988912
Change-Id: Ic41a6b194f4f5ee1f78ce9489997818dc4a2ba30