Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I rolled it back after mythria@ confirmed that it worked, looking through the logs and trying to figure out why now. I can leave the working version up for now, I'll go do that.

Ok I figured out the issue and commented on the bug. I don't think we should land this, it does fix the issue but the problem is that the /api/describe endpoint isn't auth'ing properly via simple_xhr, so when it does return you're not quite sure if it's picking up the right data. If we can fix the original test-picker cl to plumb through auth properly, we can follow up and land this.

Patch-set: 1
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>