Remove stale TODOs from qyearsley
I stopped working on the perf dashboard in 2016,
and most my TODOs are irrelevant now. Also, AFAIU,
much of the frontend code will be replaced with
V2SPA sometime soon.

This change removes the notes that I think are
likely not helpful at all now, and replaces some
if I thought there might still be something
interesting to say.

Change-Id: I71eb305344250fd55914615922151ad9db7cbb9a