Update patch set 4
Patch Set 4:

(1 comment)

Juan, PTAL at the new version.

Note that the field names in the intermediate results file are not the same as diagnostic names in histograms file (e.g. it's 'benchmark', not 'benchmarks'). Also not all values are written as lists. In those cases where multiple values don't make sense (e.g. os_name), the field name is in singular form and the value is stored as is.

My rationale behind this is to make the intermediate format self-evident, and conversion rules from it to histograms format as explicit as possible.

Patch-set: 4
Reviewer: Gerrit User 1328067 <1328067@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>