Workaround SCP issue in cros_interface
Works around a recent-ish (~January 2019) bug in SCP that can cause
valid filenames to be rejected due to strict filename checking. The
workaround is to retry the SCP command with -T if the first attempt
fails with a suitable stderr, which will cause SCP to revert to the
old behavior of not using strict filename checking.

Bug: chromium:978952
Change-Id: I5f919c0cc726d52a70527ef1c66b1093b1ecb438