Workaround SCP issue in cros_interface
Works around a recent-ish (~January 2019) bug in SCP that can cause
valid filenames to be rejected due to strict filename checking. The
workaround is to retry the SCP command with -T if the first attempt
fails with a suitable stderr, which will cause SCP to revert to the
old behavior of not using strict filename checking.

Specifically, this was occurring when attempting to pull minidumps off
of the DUT in order to symbolize them. SCP would fail with
"protocol error: filename does not match request". The exact filename
that was causing this is unknown since this only reproduces with a new
enough SCP version like the one used by the GCE bots, but is likely
related to the quoting that is done when pulling dumps, as that is
known to cause issues as per
https://unix.stackexchange.com/questions/499958/why-does-scps-strict-filename-checking-reject-quoted-last-component-but-not-oth

Bug: chromium:978952
Change-Id: I5f919c0cc726d52a70527ef1c66b1093b1ecb438