Add fake module support to device.Install API
This is the first step to moving fake module support from chromium's
apk_operations.py to device_utils.py.

Fake modules are apks that are extracted from the APKS file, but rather
than being installed like real modules, they are copied to a common
directory (in this case, /data/local/tmp/modules) so that the app can
access and emulate install at runtime. This allows apps to test their
Play Core backend. See FakeModuleInstallerBackend.java in chromium as an
example.

Bug: chromium:1000751
Change-Id: I5507ef797334804f8e67422599ada02f7ac2c385