Update patch set 5
Patch Set 5:

We discussed this offline, decided that we keep diagnostic names the same as in reserved_infos. Results Processor will support only GenericSet diagnostics so far. If we ever need to save a diagnostic of a different type, we'll decide how to represent it in an intermediate file.

Patch-set: 5
Reviewer: Gerrit User 1328067 <1328067@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>