Pinpoint: Handle all Datastore errors in migration
In this change we'd like to ignore all the errors we encounter when
migrating jobs. We also raise the batch size from 10 to 50 to quicken
the migration a little bit.

Also, we ensure that the post-get hook always runs so that we can
get the benchmark arguments in the migration.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:1006888
Change-Id: I50a690790dee2492bbf07583a65393387fa7ac6c