Pinpoint - Switch back to using google-signin
We switched to using <chops-signin> a while back, but looks like this
has some issues with logins not getting properly communicated to the
rest of the polymer app. As a result, various things can fail (like
job creation) since auth isn't properly setup. For now, switch back
to <google-signin> until these are resolved.

Bug: chromium:1007043
Change-Id: I140332d40c750201869697f8dc9b71190162d9e1