Update patch set 2
Patch Set 2: Commit-Queue+1

This is now ready for a look, folks -- apologies for the large CL!

The approach here is a little different than the other migrations, because we now have explicit dependencies being set up through factory functions and more structured configuration options. Also, we're re-using the implementation details of the quest in the evaluator implementation.

The test file is ~900 files, much of which is dict formatting and repeated parts. I suspect that can be refactored further, but chose to stop and send for review/testing instead. :)

A follow-on CL to move out the tasks from the existing quests will be forthcoming, then a CL to implement the bisection task/evaluators will come after that (at least that's the plan). :D

As always please review for readability and overall design.

Patch-set: 2
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1