Refactor all Tasks/Evaluators to new module
This change is a non-functional change to move all the task-related code
into a separate module. Along with the change in directory we've renamed
some functions that used to be 'internal' to some modules to be
'external' to be shared between quests and the tasks/evaluators that
share those implementation details.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I9ce104eca8db00a40f75e237889df81eae1e7fa5