Update patch set 3
Patch Set 3: Commit-Queue+1

Hi Folks,

This is the aforementioned 'FindCulprit' evaluator which we first implement as a performance bisection-specific evaluator. In the structure of that evaluator there are definitely pieces that could be made generic and used across A/B testing and functional testing.

So far I've covered 2/3 of the most important cases (no reproduction and speculative bisection) and I've yet to get a test for refining at changes (i.e. adding more attempts).

This is our first "native" tasks/evaluators implementation and it's beginning to show the direction we can go when we integrate this with the job creation and pubsub update handling.

As always, readability and implementation detail reviews are most appreciated. :)

Cheers

Patch-set: 3
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false